+++ spec/integration/vg_special_spec.rb
@@ -3,7 +3,7 @@ require_relative 'test_double_open_close'
 
 RSpec.describe 'VG special cases' do
   before(:context) do
-    Applitools::EyesLogger.log_handler = Logger.new(STDOUT)
+    # Applitools::EyesLogger.log_handler = Logger.new(STDOUT)
   end
   context 'Eyes Selenium SDK - Visual Grid TestDoubleOpenClose', :runner => :vg do
     include_examples 'Eyes Selenium SDK - Visual Grid TestDoubleOpenClose'